+++ R/hyper_array.R
@@ -3,7 +3,8 @@
 #' Extract the raw array data as a list of  one or more arrays. This can be the
 #' entire variable/s or after dimension-slicing using [hyper_filter()]
 #' expressions. This is a delay-breaking function and causes data to be read 
-#' from the source into raw arrays in R's native form. 
+#' from the source into raw arrays in R's native form. This list of arrays is 
+#' lightly classed as [tidync_data], with methods for [print()] and [tidync()]. 
 #'
 #' The function [hyper_array()] will act on an existing tidync object or a source
 #' string.
@@ -22,7 +23,9 @@
 #' @param force ignore caveats about large extraction and just do it
 #'
 #' @export
-#' @seealso [hyper_tbl_cube()] and [hyper_tibble()] which are also delay-breaking 
+#' @aliases tidync_data
+#' @seealso [print.tidync_data] for a description of the print summary, 
+#' [hyper_tbl_cube()] and [hyper_tibble()] which are also delay-breaking 
 #' functions that cause data to be read 
 #' @examples
 #' f <- "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc"
@@ -132,7 +135,9 @@ hyper_array.tidync <- function(x, select_var = NULL, ...,
                               dimension$count)
     }
   }
-  structure(datalist, names = varnames, transforms = transforms)
+  structure(datalist, names = varnames, 
+            transforms = transforms, 
+            source = x$source, class = "tidync_data")
 }
 #' @name hyper_array
 #' @export